+++ formatting/Viewer.format.ps1
@@ -1,0 +1,8 @@
+﻿Write-FormatView -TypeName Viewer -Property 'User Id', User -VirtualProperty @{
+    'User Id' = {
+        $($_.userId)
+    }
+    User      = {
+        "$($_.name) <$($_.login)>"
+    }
+}