+++ functions/Get-TibberConsumption.ps1
@@ -82,6 +82,7 @@
         $out = Invoke-TibberQuery @splat
 
         # Output the object
-        $out.viewer.$homeNode.consumption.nodes
+        $out = $out.viewer.$homeNode.consumption.nodes
+        $out | Add-TypeName -PSTypeName $out.__typename -PassThru
     }
 }